Code refactor with CreateNewCampaignModal.
