+++ plugins/woocommerce-admin/client/marketing/overview-multichannel/Campaigns/CreateNewCampaignModal/CreateNewCampaignModal.scss
@@ -1,0 +1,53 @@
+.woocommerce-marketing-create-campaign-modal {
+	max-width: 600px;
+
+	.woocommerce-marketing-new-campaigns {
+		padding-top: $gap-large;
+		padding-bottom: $gap;
+
+		&__question-label {
+			font-size: 14px;
+			font-weight: 600;
+			line-height: 17px;
+			margin-bottom: 16px;
+		}
+	}
+
+	.components-button.is-link {
+		text-decoration: none;
+		font-size: 14px;
+		font-weight: 600;
+		line-height: 17px;
+	}
+
+	.woocommerce_marketing_plugin_card_body {
+		padding: $gap 0;
+	}
+
+	.components-card-divider {
+		display: none;
+	}
+
+	.woocommerce-marketing-new-campaign-type {
+		padding: $gap 0;
+
+		&__name {
+			font-weight: 600;
+			font-size: 14px;
+			line-height: 20px;
+		}
+
+		&__description {
+			color: #555d66;
+		}
+
+		img,
+		svg {
+			display: block;
+		}
+	}
+
+	.woocommerce-marketing-add-channels {
+		padding-top: $gap;
+	}
+}
