+++ plugins/woocommerce-admin/client/marketing/overview-multichannel/Campaigns/CreateNewCampaignModal/index.ts
@@ -1,0 +1,1 @@
+export { CreateNewCampaignModal } from './CreateNewCampaignModal';
